feat: enhance DNP install and initialization flow with update and warning prompts
- Added support for updating existing DNP projects with missing files during initialization.
- Integrated user confirmation prompt for continuation when a project is already DNP initialized.
- Replaced errors with warnings for smoother user interaction during initialization.
- Updated DNP structure creation with improved error handling and placeholder substitutions.
- Update tests

Issue NMO-656

+++ src/lib/commands/init.bash
@@ -27,6 +27,23 @@ test -d "${DNP_ROOT:?err}" || { echo -e "${dnp_error_prefix} librairy load error
 test -d "${DNP_LIB_PATH:?err}" || { echo -e "${dnp_error_prefix} librairy load error!" ; exit 1 ; }
 
 # ::::Command functions::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+
+#function n2st::seek_and_modify_string_in_file() {
+#  # Note: Temporary ovverride of the original (see task N2ST-52)
+#  local TMP_SEEK="${1}"
+#  local TMP_CHANGE_FOR="${2}"
+#  local TMP_FILE_PATH="${3}"
+#
+#  # Note:
+#  #   - Character ';' is used as a delimiter
+#  #   - Keep -i flag for portability to Mac OsX (it's analogue to --in-place flag)
+#  #   - .bak is the backup extension convention and is required by -i
+#  sudo sed -i.bak "s;${TMP_SEEK};${TMP_CHANGE_FOR};" "${TMP_FILE_PATH}" || return 1
+#  sudo rm "${TMP_FILE_PATH}.bak" || return 1
+#  return 0
+#}
+
+
 function dnp::get_acronym() {
     # Find the acronym of a dash or underscrore sepatared multi-word name.
     local super_project_name="$1"
@@ -85,60 +102,87 @@ function dnp::init_command() {
     local project_git_remote_url
     local super_project_name
     local super_project_user
-    project_git_remote_url="$( git remote get-url origin )"
-    super_project_name="$( basename "${project_git_remote_url}" .git )}"
+    project_git_remote_url="$( git remote get-url origin )" || return 1
+    super_project_name="$( basename "${project_git_remote_url}" .git )"
     super_project_user="$(id -un)"
 
     # Check if .dockerized_norlab_project already exists
     if [[ -d ".dockerized_norlab_project" ]]; then
-        n2st::print_msg_error_and_exit "${MSG_DIMMED_FORMAT}.dockerized_norlab_project${MSG_END_FORMAT} directory already exists.\nThis project is already DNP initialized."
+        n2st::print_msg_warning "This project is already DNP initialized since ${MSG_DIMMED_FORMAT}.dockerized_norlab_project${MSG_END_FORMAT} directory already exists.\nIf you continue, missing file will be created and existing one will be updated."
+        read -r -n 1 -p "Do you want to continue [y/N]" option_update
+        if [[ "${option_update}" == "y" || "${option_update}" == "Y" ]]; then
+          :
+        else
+          n2st::print_msg "See you"
+          return 0
+        fi
     fi
 
     # ====Begin====================================================================================
     n2st::print_msg "Initializing DNP project: ${super_project_name} in ${super_project_root}"
 
-    # Create the .dockerized_norlab_project directory
-    sudo mkdir -p .dockerized_norlab_project
+#    # Create the .dockerized_norlab_project directory
+#    sudo mkdir -p .dockerized_norlab_project || return 1
 
     # Copy template files
-    cp -r "${DNP_LIB_PATH}/template/.dockerized_norlab_project/"* .dockerized_norlab_project/
+    sudo cp --update -r "${DNP_LIB_PATH}/template/.dockerized_norlab_project/" . || return 1
 
-    cd "${super_project_root}/.dockerized_norlab_project/" || exit 1
+    cd "${super_project_root}/.dockerized_norlab_project/" || return 1
 
     # Rename the super project DNP meta .env file
-    cp ".env.PLACEHOLDER_SUPER_PROJECT_NAME" ".env.${super_project_name}"
+    sudo mv --force "${super_project_root}/.dockerized_norlab_project/.env.PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_root}/.dockerized_norlab_project/.env.${super_project_name}" || return 1
 
     # Replace placeholders in the .env.dnp file
-    cd "${super_project_root}/.dockerized_norlab_project/configuration/" || exit 1
+    cd "${super_project_root}/.dockerized_norlab_project/configuration/" || return 1
     local super_project_acronym
     super_project_acronym="$(dnp::get_super_project_acronym "${super_project_name}")"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_PROJECT_GIT_REMOTE_URL" "${project_git_remote_url}" ".env.dnp"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_CONTAINER_NAME" "IamDNP_${super_project_acronym}" ".env.dnp"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_PROJECT_ALIAS_PREFIX" "${super_project_acronym}" ".env.dnp"
+
+    {
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_PROJECT_GIT_REMOTE_URL" "${project_git_remote_url}" ".env.dnp" &&
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_CONTAINER_NAME" "IamDNP_${super_project_acronym}" ".env.dnp" &&
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_PROJECT_ALIAS_PREFIX" "${super_project_acronym}" ".env.dnp"
+    } || return 1
 
     # Replace placeholders in the DNP readme file
-    cd "${super_project_root}/.dockerized_norlab_project/" || exit 1
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_CONTAINER_NAME" "IamDNP_${super_project_acronym}" "README.md"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_USER" "${super_project_user}" "README.md"
+    cd "${super_project_root}/.dockerized_norlab_project/" || return 1
+    {
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_DN_CONTAINER_NAME" "IamDNP_${super_project_acronym}" "README.md" &&
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_USER" "${super_project_user}" "README.md"
+    }  || return 1
 
     # Note: There is three ocurence
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md"
-    n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md"
+    {
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md" &&
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md" &&
+      n2st::seek_and_modify_string_in_file "PLACEHOLDER_SUPER_PROJECT_NAME" "${super_project_name}" "README.md"
+    } || return 1
 
     # ....Create root repository required directories..............................................
-    cd "${super_project_root}" || exit 1
+    cd "${super_project_root}" || return 1
 
-    sudo mkdir -p artifact
-    sudo mkdir -p external_data
-    sudo mkdir -p src/launcher
-    sudo mkdir -p src/tools
-    sudo mkdir -p tests
+    {
+      sudo mkdir -p artifact/optuna_storage &&
+      sudo mkdir -p external_data &&
+      sudo mkdir -p src/launcher/configs &&
+      sudo mkdir -p src/tools &&
+      sudo mkdir -p tests
+    } || return 1
+
+    sudo cp --update -r "${DNP_LIB_PATH}/template/artifact/README.md" artifact/ || return 1
+    sudo cp --update -r "${DNP_LIB_PATH}/template/artifact/optuna_storage/README.md" artifact/optuna_storage/ || return 1
+    sudo cp --update -r "${DNP_LIB_PATH}/template/external_data/README.md" external_data/ || return 1
+    sudo cp --update -r "${DNP_LIB_PATH}/template/src/launcher" src/ || return 1
+    sudo cp --update -r "${DNP_LIB_PATH}/template/src/tools" src/ || return 1
+    if [[ ! -f "src/README.md" ]]; then
+      sudo cp --update -r "${DNP_LIB_PATH}/template/src/README.md" src/ || return 1
+    fi
+    sudo cp --update -r "${DNP_LIB_PATH}/template/tests" . || return 1
 
     # ....Create root README.md files if it does't exist...........................................
-    cd "${super_project_root}" || exit 1
+    cd "${super_project_root}" || return 1
 
     if [[ ! -f "README.md" ]]; then
+        sudo touch "README.md"  || return 1
         cat > "README.md" << EOF
 # ${super_project_name}
 
@@ -153,7 +197,7 @@ EOF
 
     if [[ ! -f ".gitignore" ]]; then
         # Case: file does not exist => copy template
-        cp "${DNP_LIB_PATH}/template/.gitignore" .gitignore
+        sudo cp "${DNP_LIB_PATH}/template/.gitignore" .gitignore || return 1
     else
         # Case: file exist => append required .gitignore entries
         cat >> ".gitignore" << EOF
@@ -175,7 +219,7 @@ EOF
 
     if [[ ! -f ".dockerignore" ]]; then
         # Case: file does not exist => copy template
-        cp "${DNP_LIB_PATH}/template/.dockerignore" .dockerignore
+        sudo cp "${DNP_LIB_PATH}/template/.dockerignore" .dockerignore  || return 1
     else
         # Case: file exist => append required .gitignore entries
       cat >> ".dockerignore" << EOF
@@ -201,14 +245,16 @@ EOF
     source "${DNP_LIB_PATH}/core/utils/setup_host_for_running_this_super_project.bash" || return 1
 
     # ====Teardown=================================================================================
+    tree -l --dirsfirst -a -I visual -I .git -L 7 "$PWD"
+
     n2st::print_msg "DNP project initialized successfully.
 You can now use ${MSG_DIMMED_FORMAT}dnp${MSG_END_FORMAT} to manage your project.
 To get started:
   1. Execute ${MSG_DIMMED_FORMAT}dnp help${MSG_END_FORMAT} to see available command
-  2. Read instruction in '.dockerized_norlab_project/README.md'
-  3. Check documentation at https://github.com/norlab-ulaval/dockerized-norlab-project.git.
+  2. Read instruction in ${MSG_DIMMED_FORMAT}${super_project_name}/.dockerized_norlab_project/README.md${MSG_END_FORMAT}
+  3. Check documentation at https://github.com/norlab-ulaval/dockerized-norlab-project
   4. Stay awesome
 "
-    cd "$super_project_root" || exit 1
+    cd "$super_project_root" || return 1
     return 0
 }
